+++ NEWS.md
@@ -6,6 +6,11 @@
 
 * Add `ADLB` data set. (#450)
 
+* `shuffle_ard()` fills overall or group statistics with `"Overall <column_name>"` or `"Any <column_name>"`. (#337, @dragosmg)
+
+* `shuffle_ard()` messages if `"Overall <column_names>"` is accidentally present in the data and creates a unique label. (#465, @dragosmg)
+
+
 # cards 0.6.1
 
 ## New Features and Functions